requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
numpy==1.26.4
tensorflow-cpu==2.18.0
tensorflow-hub==0.16.1
pymongo==4.6.1
fastapi==0.112.0
uvicorn==0.29.0

# For local development (not used in Docker):
# pycairo>=1.16.0
# PyGObject>=3.40.0